move dmme.common into dmme namespace
